Module Docker

Interface to Docker Remote API.

See the Docker API if needed to complement the documentation below.

version 0.2.2

val set_default_addr : Unix.sockaddr -> unit

Set the address the Docker daemon listens to. This will be used by all funtions of this API unless the optional parameter ~addr is used.

exception No_such_container of string

No_such_container id is raised to notify that the container id does not exist.

exception Failure of string * string

Failure(fn_name, msg) is raised when the requested action fails to be performed (because, say, the container stopped,...). fn_name is the name of the function raising the error and msg an explanation.

exception Invalid_argument of string

Invalid_argument msg can be raised by any function when an incorrect argument. msg is the function raising the exception.

exception Server_error of string

Server_error fn_name indicates that the server encountered an error or that the returned response is incorrect. fn_name is the function raising the error.

exception Error of string * string

Error(fn_name, msg) is raised for connection of protocol errors. fn_name is the function raising the error and msg is a possible explanation for the error. Typically, this exception should not be raised with a well behaving server.

module Stream : sig ... end

A stream returned by some Docker functions.

module Container : sig ... end

module Image : sig ... end

type version = {

api_version : string;

version : string;

git_commit : string;

go_version : string;

}

val version : ?addr:Unix.sockaddr -> unit -> version